refactor: raceAbort を @vicissitude/shared/functions に集約 (#654)
Issue #654 の対応:

- `raceAbort` を `packages/agent/src/runner.ts` から `packages/shared/src/functions.ts`
  に移動し、他パッケージから再利用可能にする。
- `abortReasonToError` は `raceAbort` 内部専用ヘルパーとして shared 内に閉じる
  （module private、export しない）。
- `withTimeout` と `raceAbort` の使い分けポリシーを functions.ts のコメントに明記:
  signal 伝播不要 → `withTimeout`、signal 伝播要 → `raceAbort`。

`withTimeout` は setTimeout ベースで AbortSignal を扱えないため統合はせず、
用途別に併存させる方針。
